+++ GeometricObjects/Primitives/Plane.h
@@ -16,7 +16,7 @@ public:
   virtual ~Plane(void);                         				// destructor
 
   virtual bool hit(const Ray& ray, double& tmin, ShadeRec& sr) const;
-  virtual bool shadow_hit(const Ray& ray, double& tmin) const;
+  virtual bool shadow_hit(const Ray& ray, float& tmin) const;
 
 private:
   Point3D 	a;   				// point through which plane passes
